Remove BehaviorFromToLinkArgs
This type was meant to translate a function parameters into an object,
but this turns out not to be possible in typescript. Function parameters
are represented as a named tuple, whose names are known by typescript
for documentation purposes, but not exposed to the type system.

see https://stackoverflow.com/a/69713319